Update libs. Fix process protobuf generated enums. Fix generator for ApiResponse with reference. Fix code style errors. Get version of `jackson-dataformat-yaml` from micronaut versions catalog

+++ test-suite-java-server-generator/src/main/java/io/micronaut/openapi/test/api/ResponseBodyController.java
@@ -74,7 +74,7 @@ public class ResponseBodyController implements ResponseBodyApi {
 
     @Override
     public Mono<Page<SimpleModel>> getPaginatedSimpleModel(Pageable pageable) {
-        return Mono.just(Page.of(SIMPLE_MODELS, pageable, SIMPLE_MODELS.size()));
+        return Mono.just(Page.of(SIMPLE_MODELS, pageable, (long) SIMPLE_MODELS.size()));
     }
 
     @Override
